requirements.txt
asgiref==3.6.0
Babel==2.12.1
beautifulsoup4==4.12.2
calendra==7.6.1
certifi==2023.5.7
charset-normalizer==3.1.0
convertdate==2.4.0
cssselect2==0.7.0
defusedxml==0.7.1
diff-match-patch==20230430
distlib==0.3.6
Django==4.1.5
django-bootstrap4==23.1
django-calendarium==1.3.4
django-cleanup==7.0.0
django-filer==3.0.0
django-filter==23.2
django-import-export==3.2.0
django-ipware==5.0.0
django-libs==2.0.3
django-phonenumber-field==7.0.2
django-polymorphic==3.1.0
easy-thumbnails==2.8.5
et-xmlfile==1.1.0
filelock==3.9.0
geographiclib==2.0
geopy==2.3.0
idna==3.4
import-export==0.3.1
lunardate==0.2.0
lxml==4.9.3
MarkupPy==1.14
more-itertools==9.1.0
numpy==1.24.2
odfpy==1.4.1
openpyxl==3.1.2
pandas==2.0.1
phonenumbers==8.13.10
pi==0.1.2
Pillow==9.5.0
platformdirs==2.6.2
pyluach==2.2.0
PyMeeus==0.5.12
pyserial==3.5
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
reportlab==4.0.4
requests==2.31.0
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.3
svglib==1.5.1
tablib==3.4.0
tinycss2==1.2.1
tzdata==2022.7
upgrade-pip==0.1.4
urllib3==2.0.3
virtualenv==20.17.1
webencodings==0.5.1
xlrd==2.0.1
xlwt==1.3.0
